feat: add SEO metadata to pages
