+++ app/legal/disclaimer/page.tsx
@@ -1,3 +1,27 @@
+import type { Metadata } from 'next';
+
+const url = 'https://lexlens.ai/legal/disclaimer';
+
+export const metadata: Metadata = {
+  title: 'Disclaimer – LexLens',
+  description: 'Important disclaimers about using LexLens. Informational only; not legal advice.',
+  alternates: {
+    canonical: url,
+  },
+  openGraph: {
+    title: 'Disclaimer – LexLens',
+    description: 'Important disclaimers about using LexLens. Informational only; not legal advice.',
+    url,
+    siteName: 'LexLens',
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary',
+    title: 'Disclaimer – LexLens',
+    description: 'Important disclaimers about using LexLens. Informational only; not legal advice.',
+  },
+};
+
 export default function DisclaimerPage() {
   return (
     <div className="prose max-w-3xl">
@@ -6,5 +30,5 @@ export default function DisclaimerPage() {
       <p><strong>No confidentiality.</strong> Do not share sensitive personal data or case details. Using this service does not create attorney–client privilege.</p>
       <p><strong>Jurisdiction variance.</strong> Laws differ by state and change over time. Content may not reflect the latest updates. Always verify via linked sources.</p>
     </div>
-  )
+  );
 }
